+++ src/clashing_inputs.rs
@@ -8,8 +8,8 @@ use crate::user_input::{InputKind, UserInput};
 use crate::Actionlike;
 
 use bevy::prelude::Resource;
+use bevy::utils::HashSet;
 use itertools::Itertools;
-use petitset::PetitSet;
 use serde::{Deserialize, Serialize};
 use std::cmp::Ordering;
 use std::marker::PhantomData;
@@ -154,8 +154,8 @@ impl<A: Actionlike> InputMap<A> {
     fn possible_clash(&self, action_a: A, action_b: A) -> Option<Clash<A>> {
         let mut clash = Clash::new(action_a.clone(), action_b.clone());
 
-        for input_a in self.get(action_a).iter() {
-            for input_b in self.get(action_b.clone()).iter() {
+        for input_a in self.get(action_a)? {
+            for input_b in self.get(action_b)? {
                 if input_a.clashes(input_b) {
                     clash.inputs_a.push(input_a.clone());
                     clash.inputs_b.push(input_b.clone());
@@ -212,7 +212,7 @@ impl<A: Actionlike> Clash<A> {
 
 // Does the `button` clash with the `chord`?
 #[must_use]
-fn button_chord_clash(button: &InputKind, chord: &PetitSet<InputKind, 8>) -> bool {
+fn button_chord_clash(button: &InputKind, chord: &HashSet<InputKind>) -> bool {
     if chord.len() <= 1 {
         return false;
     }
@@ -222,7 +222,7 @@ fn button_chord_clash(button: &InputKind, chord: &PetitSet<InputKind, 8>) -> boo
 
 // Does the `dpad` clash with the `chord`?
 #[must_use]
-fn dpad_chord_clash(dpad: &VirtualDPad, chord: &PetitSet<InputKind, 8>) -> bool {
+fn dpad_chord_clash(dpad: &VirtualDPad, chord: &HashSet<InputKind>) -> bool {
     if chord.len() <= 1 {
         return false;
     }
@@ -275,7 +275,7 @@ fn virtual_axis_dpad_clash(axis: &VirtualAxis, dpad: &VirtualDPad) -> bool {
 }
 
 #[must_use]
-fn virtual_axis_chord_clash(axis: &VirtualAxis, chord: &PetitSet<InputKind, 8>) -> bool {
+fn virtual_axis_chord_clash(axis: &VirtualAxis, chord: &HashSet<InputKind>) -> bool {
     if chord.len() <= 1 {
         return false;
     }
@@ -293,7 +293,7 @@ fn virtual_axis_virtual_axis_clash(axis1: &VirtualAxis, axis2: &VirtualAxis) -> 
 
 /// Does the `chord_a` clash with `chord_b`?
 #[must_use]
-fn chord_chord_clash(chord_a: &PetitSet<InputKind, 8>, chord_b: &PetitSet<InputKind, 8>) -> bool {
+fn chord_chord_clash(chord_a: &HashSet<InputKind>, chord_b: &HashSet<InputKind>) -> bool {
     if chord_a.len() <= 1 || chord_b.len() <= 1 {
         return false;
     }